# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для микро сервиса",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apiros/api",
            "description": "Основной API"
        }
    ],
    "paths": {
        "/jobs": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Получение списка jobs",
                "description": "Метод возвращает данные ...",
                "operationId": "getJobs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/task": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Получение списка задач",
                "description": "Метод возвращает данные ...",
                "operationId": "getTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id task",
                        "schema": {
                            "type": "sting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Создание данных",
                "description": "Метод содает данные и шифрует их ...",
                "operationId": "setTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/task/stop/{id}": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Получение списка задач",
                "description": "Метод возвращает данные ...",
                "operationId": "stoptask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id task",
                        "schema": {
                            "type": "sting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/task/start/{id}": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Получение списка задач",
                "description": "Метод возвращает данные ...",
                "operationId": "starttask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id task",
                        "schema": {
                            "type": "sting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/task/pack": {
            "post": {
                "tags": [
                    "task pack"
                ],
                "summary": "Создание данных",
                "description": "Метод содает данные пакетом и шифрует их ...",
                "operationId": "setPackTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskPackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/task/pack/all": {
            "post": {
                "tags": [
                    "task pack"
                ],
                "summary": "Создание данных",
                "description": "Возвращает пакет данных",
                "operationId": "GetPackTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskPackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/task/pack/stop": {
            "post": {
                "tags": [
                    "task pack"
                ],
                "summary": "Создание данных",
                "description": "Останавливает шифрование пакетно",
                "operationId": "stopPackTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskPackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/task/pack/start": {
            "post": {
                "tags": [
                    "task pack"
                ],
                "summary": "Создание данных",
                "description": "Запускает шифрование пакетно",
                "operationId": "startPackTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskPackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TaskPackRequest": {
                "title": "TaskPackRequest",
                "description": "Task Pack Request",
                "properties": {
                    "packege": {
                        "title": "Packege",
                        "description": "Массив обьектов для шифрования",
                        "type": "string",
                        "example": "[{\"value\":\"qwert\",\"repetitions\":12}]"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskPackRequest"
                }
            },
            "TaskRequest": {
                "title": "TaskRequest",
                "description": "Task Request",
                "properties": {
                    "value": {
                        "title": "Value",
                        "description": "строка которую шифрует",
                        "type": "string",
                        "example": "qwerty"
                    },
                    "repetitions": {
                        "title": "Repetitions",
                        "description": "количетсво проходов шифра",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskRequest"
                }
            },
            "Task": {
                "title": "Task",
                "description": "Task model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "example": 1
                    },
                    "value": {
                        "title": "Value",
                        "description": "строка которую шифрует",
                        "type": "string",
                        "example": "qwerty"
                    },
                    "repetitions": {
                        "title": "Repetitions",
                        "description": "количетсво проходов шифра",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "repetition_repetitions": {
                        "title": "Repetition Repetitions",
                        "description": "остаток проходов шифра",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Task"
                }
            }
        }
    }
}